I've added: Forage success, development rate I've deleted: Inner time
The backwards plot doesn't look good (or at least as the previous one), and also almost always it's always better to invest in growth than in performance. The invest plot looks weird, they always grow to their maximum size.
The survival plot doesn't work well and I don't know why, I have to revise the Population matrix and how it is created during the iteration.

This is the most "realistic" model, but it doesn't behave like what we see in nature. :/